Update - adding new methods

+++ src/src/bg/tu_varna/sit/Line.java
@@ -5,10 +5,10 @@ import java.awt.geom.Line2D;
 
 public class Line extends Figure {
 
-    private int xEnd;
-    private int yEnd;
+    private double xEnd;
+    private double yEnd;
 
-    public Line(String color, double xStart, double yStart, int xEnd, int yEnd) {
+    public Line(String color, double xStart, double yStart, double xEnd, double yEnd) {
         super(color,xStart,yStart);
         this.xEnd = xEnd;
         this.yEnd = yEnd;
